Deployment on npmjs.com
* Use webpack + babel w/ ES2015 presets
* Modularize HTML Highlighter library
* Convert library source code to ES2015
