+++ examples/binary-search/binary-search.html
