+++ examples/binary-search/binary-search.js
