+++ package.json
@@ -1,20 +1,47 @@
 {
   "name": "html-highlighter",
+  "title": "Html Highlighter",
+  "description": "",
   "version": "0.1.0",
-  "description": "HTML Highlighter",
+  "main": "dist/html_highlighter.min.js",
   "repository": {
     "type": "git",
-    "url": "git@github.com:dossier/html-highlighter.git"
+    "url": "git+https://github.com/dossier/html-highlighter.git"
   },
-  "dependencies": { },
+  "keywords": [
+    "html",
+    "highlighter"
+  ],
+  "author": "Diffeo",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/dossier/html-highlighter.git/issues"
+  },
+  "peerDependencies": {
+    "jquery": "~2.2.4"
+  },
+  "dependencies": {},
   "devDependencies": {
-    "grunt": "0.4.x",
-    "grunt-cli": "0.1.x",
-    "grunt-saucelabs": "8.x",
-    "grunt-contrib-connect": "0.x",
-    "grunt-contrib-watch": "0.x"
+    "babel-core": "^6.13.2",
+    "babel-eslint": "^6.1.2",
+    "babel-loader": "^6.2.5",
+    "babel-preset-es2015": "^6.13.2",
+    "documentation-loader": "^0.1.1-beta",
+    "eslint": "^3.3.1",
+    "eslint-loader": "^1.5.0",
+    "file-loader": "^0.9.0",
+    "grunt": "~1.0.1",
+    "grunt-cli": "~1.2.0",
+    "grunt-contrib-connect": "~1.0.2",
+    "grunt-contrib-watch": "~1.0.0",
+    "grunt-saucelabs": "~9.0.0",
+    "http-server": "^0.9.0",
+    "webpack": "^1.13.2",
+    "webpack-dev-server": "^1.15.0"
   },
   "scripts": {
+    "prebuild": "npm install",
+    "build": "webpack --progress --color",
     "test": "grunt test"
   }
 }
