+++ src/main.js
@@ -1,0 +1,13 @@
+import HtmlHighlighter from "./htmlhighlighter.js";
+import Ui from "./ui/ui.js";
+import RangeHighlighter from "./rangehighlighter.js";
+import TextFinder from "./textfinder.js";
+import XPathFinder from "./xpathfinder.js";
+
+export {
+  HtmlHighlighter,
+  Ui,
+  RangeHighlighter,
+  TextFinder,
+  XPathFinder
+};
