+++ test/src
@@ -1,1 +1,1 @@
-../src
+../dist
